:recycle: Move transactionDetailView to singleTransactionV2 folder
