+++ src/components/singleTransactionV2/transactionDetailView/transactionDetailViewV2.css
