+++ src/components/singleTransactionV2/transactionDetailView/transactionVotes.js
