+++ src/components/singleTransactionV2/transactionDetailView/transactionVotes.test.js
